Implement the two-stage gathering of Hit data into Matriplex.
This is super dirty as it relies on the fact that sizeof(Hit)=64.

ChewIn() -- do both stages.

Contaginate() -- read hits into contiguous memory.

Plexify() -- fill Matriplex from contaginated memory.

The catch here is that we can call Contaginate() once for whole Hits
and then call Plexify() for error matrix and for parameters.