Port of clone-engine and propagation-at-end to devel.
This is not perfect yet but all features made it across.

List of issues is in comment in mkFit/buildtestMPlex.h.

Something is wrong with FindTracksBestHit() ... didn't have time to
check but as no tracks are found it could be something silly (I was
moving things around and probably missed some variable name or
location change).