CMS-2017 tracking in barrel works.
`./mkFit --read --file-name /tmp/mu_brl-100-10.bin --build-bh --geom CMS-2017 --input-version 0 --cmssw-seeds --num-events 1`

- Cleanup handling and preparation of seeds. Simplifies MkBuilder and
  functions in buildtestMPlex.cc.
    - add top-level function MkBuilder::PrepareSeeds().
      - add internal function MkBuilder::import_seeds() and a helper function for
          sorting of seeds based on eta. This is used both for:
              - importing seeds from a file;
                  - and importing seeds from MC tracks.

- Event::read_in() add define statements for what parts to dump.

- Track:
  - Fix bug in getLastGoodHitIdx/Lyr().
    - Rename nGoodHitIdx_ to nGoodHits :)

- mkFit/mkFit.cc: --num-events X is also honoured when --read is in force, but
  only if it is *after* the --read argument.
    *This was actually in previous commit, forgot to document it.*

*Remains to be done:*

Layer where we start searching for hits is set from SteeringParams. This is
still fine for barrel but will be hard for CMS endcap as seeds come in with
different last layer! So I'll have to be inventive there and on each layer
select candidates to propagate further. This will be similar to what needs to
be done for transition region.

This also does not support different iterations, for this we'd need something
like IterationParams.