Introduce MkFinder - a splice off MkFitter.
MkFitter was becoming unwieldy plus there was the assumption of all tracks
having the same number of hits which is not true in endcap building.

This is a partial conversion and there is some repetiotion of the code.

BestHit finding already uses the MkFinder, Std and CE use the old code.