Remove cut for minimum number of hits in seed
This commit is mainly to work around a "feature" of CMSSW EDProducer
that converts pixel reco::Tracks to TrajectorySeeds that ignores the
fourth hit of the seed.

On ther other hand, any seed selection (that depends purely on the
seed properties) should really be done by an independent producer in
CMSSW (or by non-CMSSW-called code within the standalone application)
as that would be much more flexible.