Combine backward fit / backward search.
Before backward fit stash away seed-hits and re-inject them after SelectHitIndices
is called on that layer.

If this works the code in MkBuilder should be put into functions.

In backward search scale down chi2 of hits that were present in the seed.

Do not add invalid hit when a seed hit is taken from stash.